+++ toro/__init__.py
@@ -10,7 +10,7 @@ from tornado import stack_context
 from tornado.ioloop import IOLoop
 
 
-version_tuple = (0, 2)
+version_tuple = (0, 3)
 
 version = '.'.join(map(str, version_tuple))
 """Current version of Toro."""